Fixed plugin not working on ios 13 with xcode 11 build
Due to the changes in ios13 to refine the multi-app experience for iOS and iPadOS, we're now required to help the OS out when it's making the decision on where and how to display the OAuth Alert and Web login flow. We now let the ASWebAuthenticationSession know which window is presenting the OAuth request by implementing ASWebAuthenticationPresentationContextProviding interface in your presenting View Controller which returns the relevant window in the presentationAnchor method.

I don't see how we can override the FlutterViewcontroller by default to implement this new Interface.

This however, solves the runtime issue when building with xcode11 that uses the stable SF for ios13.