🎉 Add support for Flutter "add to app" (#106)
* The current implementation assumes the plugin is used in a pure Flutter app. It casts the app's rootViewController to FlutterViewController.

This supports using this plugin in an "add to app" use case. The implementation does not assume the rootViewController is an instance of FlutterViewController, but rather starts with the presented view controller, and navigates up the stack to the root controller. It also handles in the case of a non-modal screen, in which case it would be a UINavigationController. We could probably just use the presented view controller, rather than navigating up the stack to the root to use as the session context provider.

* Use guarded cast instead of forced unwrap.

* fix(swift): make changes from code review

* fix(swift): eliminate !s and unused code

Co-authored-by: junying1 <jying@simplec.com>